pyproject.toml
[project]
name = "code-war"
version = "0.1.0"
description = ""
authors = [
    {name = "pedrowowbr",email = "edupedrobraga@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0" 

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = {version = "^0.115.2", extras = ["standard"]}
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.1.1"
alembic = "^1.16.2"
streamlit = "^1.47.0"
plotly = "^6.2.0"
pandas = "^2.2.2"

[tool.poetry]
packages = [{include = "code_war", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.1"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
alembic = "^1.16.4"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/code_war/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/code_war -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"